bringup: upgrade kernel from 6.11 to 6.12 [1/1]
PD#SWPL-188184

Problem:
upgrade kernel from 6.11 to 6.12

Solution:
AML_KERNEL_VERSION	3
UPSTREAM_VERSION	"5a81cc3248a2"
AML_PATCH_VERSION	"5a81cc3248a2"
MERGE_DATE		"2024-10-09"
1. modify build script
2. port auto patch and move common-mainline to common16-6.12
3. modify the code due to the change of
the remove output_poll_changed in struct drm_mode_config_funcs
4. modify the code due to the change of the remove GPIOF_DIR_OUT
5. modify the code due to the change of the remove no_llseek
6. modify the code due to the change of the struct fd
7. modify the code due to the change of the struct mem_cgroup_reclaim_iter
8. modify the code due to the change of the file unaligned.h

Verify:
local

Test:
bootup to console in fatload mode

+++ amlogic_utils.sh
@@ -894,7 +894,12 @@ function modules_install() {
 	popd
 
 	if [[ ${BAZEL} == "1" ]]; then
-		cp ${DIST_DIR}/vmlinux ${OUT_AMLOGIC_DIR}/symbols
+		if [[ -f ${DIST_DIR}/vmlinux ]]; then
+			cp ${DIST_DIR}/vmlinux ${OUT_AMLOGIC_DIR}/symbols
+		else
+			vmlinux_file=`find ${BAZEL_OUT} -name vmlinux | grep kernel_aarch64`
+			cp ${vmlinux_file} ${OUT_AMLOGIC_DIR}/symbols
+		fi
 
 		find ${BAZEL_OUT} -name *.ko | grep "unstripped" | while read module; do
 		        cp ${module} ${OUT_AMLOGIC_DIR}/symbols
